Move server/config/constants > config/dbTypes
