+++ server/db/sequelize/sequelize_config.js
@@ -1,5 +1,5 @@
-const DB_TYPE = require('../../config/appConfig').DB_TYPE;
-const DB_TYPES = require('../../config/constants').DB_TYPES;
+import { DB_TYPE } from '../../config/appConfig';
+import { DB_TYPES } from '../../../config/dbTypes';
 
 switch (DB_TYPE) {
   case DB_TYPES.POSTGRES:
